NotificationService
- apply new learning to NotificationService
- move implementation to own package
