+++ services/src/main/java/knaufdan/android/services/userinteraction/notification/INotificationService.kt
@@ -3,7 +3,6 @@ package knaufdan.android.services.userinteraction.notification
 import kotlin.reflect.KClass
 
 interface INotificationService {
-
     fun configure(adjust: INotificationServiceConfig.() -> Unit)
 
     fun sendNotification(
