+++ services/src/main/java/knaufdan/android/services/userinteraction/notification/INotificationServiceConfig.kt
@@ -1,7 +1,6 @@
 package knaufdan.android.services.userinteraction.notification
 
 interface INotificationServiceConfig {
-
     fun setNotificationChannel(
         channelId: String,
         channelName: String = "",
